Fix x64 warning with casts.
This reverts that change to decompileHLSL.h, so that it doesn't have to use those awful Microsoft style headers, and can stay more normal C++.

Instead of SIZE_T for the decompiledSize, revert to long.

This requires a cast to long when used by the callers.

+++ HLSLDecompiler/DecompileHLSL.h
@@ -8,7 +8,7 @@ struct ParseParameters
 {
 	const void *bytecode;
 	const char *decompiled;
-	SIZE_T decompiledSize;
+	long decompiledSize;
 
 	bool fixSvPosition;
 	bool fixLightPosition;
